move new tests to a new file
